feat: added validators with constructor injection and validated request model

+++ src/CleanArch.Application/Commands/Note/DeleteNote/DeleteNoteCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using CleanArch.Application.Interfaces.UnitOfWork;
+using CleanArch.Application.Validations.Note;
 using CleanArch.Domain.Common;
 using CleanArch.Domain.Constants;
+using FluentValidation;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -9,13 +11,19 @@ namespace CleanArch.Application.Commands.Note.DeleteNote
 {
     public class DeleteNoteCommandHandler : IRequestHandler<DeleteNoteCommandRequest, AppResponse>
     {
-        private IUnitOfWork UoW;
-        public DeleteNoteCommandHandler(IUnitOfWork uoW)
+        private readonly IUnitOfWork UoW;
+        private readonly DeleteNoteValidator validator;
+        public DeleteNoteCommandHandler(IUnitOfWork uoW, DeleteNoteValidator validator)
         {
             UoW = uoW;
+            this.validator = validator;
         }
         public async Task<AppResponse> Handle(DeleteNoteCommandRequest request, CancellationToken cancellationToken)
         {
+            var validationResult = validator.Validate(request);
+            if (validationResult.Errors.Count > 0)
+                return new ErrorResponse(Messages.VALIDATION_ERROR);
+
             var result = await UoW.NoteRepository.RemoveAsync(request.Id);
             await UoW.SaveAsync();
             return result == null
