+++ src/CleanArch.Application/Commands/Tag/DeleteTag/DeleteTagCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using CleanArch.Application.Interfaces.UnitOfWork;
+using CleanArch.Application.Validations.Tag;
 using CleanArch.Domain.Common;
 using CleanArch.Domain.Constants;
 using MediatR;
@@ -11,15 +12,19 @@ namespace CleanArch.Application.Commands.Tag.DeleteTag
     public class DeleteTagCommandHandler : IRequestHandler<DeleteTagCommandRequest, AppResponse>
     {
         private readonly IUnitOfWork UoW;
-        private readonly IMapper mapper;
-        public DeleteTagCommandHandler(IMapper mapper, IUnitOfWork UoW)
+        private readonly DeleteTagValidator validator;
+        public DeleteTagCommandHandler(IUnitOfWork UoW, DeleteTagValidator validator)
         {
             this.UoW = UoW;
-            this.mapper = mapper;
+            this.validator = validator;
         }
 
         public async Task<AppResponse> Handle(DeleteTagCommandRequest request, CancellationToken cancellationToken)
         {
+            var validationResult = validator.Validate(request);
+            if (validationResult.Errors.Count > 0)
+                return new ErrorResponse(Messages.VALIDATION_ERROR);
+
             var result = await UoW.TagRepository.RemoveAsync(request.Id);
             await UoW.SaveAsync();
             return result == null
